# attached_assets/OtherPoker Files/stitch-poker/lib/api-zod/src/generated/types/playerState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stitch Poker API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerStateStatus } from "./playerStateStatus";

export interface PlayerState {
  seatIndex: number;
  username: string;
  displayName: string;
  avatarUrl?: string;
  chips: number;
  bet: number;
  status: PlayerStateStatus;
  isDealer: boolean;
  isTurn: boolean;
}