# attached_assets/OtherPoker Files/stitch-poker/lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stitch Poker API
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentGameType } from "./tournamentGameType";
import type { TournamentStatus } from "./tournamentStatus";

export interface Tournament {
  id: number;
  name: string;
  description?: string;
  gameType: TournamentGameType;
  buyIn: number;
  prizePool?: number;
  maxPlayers: number;
  registeredPlayers: number;
  startTime?: Date;
  status: TournamentStatus;
  blindStructure?: string;
  startingChips?: number;
  levels?: number;
}